Features/VNFramework/broker-scaffold/requirements.txt
# VNF Broker - Python Dependencies

fastapi>=0.104.0
pydantic>=2.5.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
PyJWT>=2.8.0
python-multipart>=0.0.6
pyyaml>=6.0.1
jinja2>=3.1.2
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx-mock>=0.10.0
